pyproject.toml
[tool.poetry]
name = "pyspark-streaming-poc"
version = "0.0.1"
description = "POC for data streaming using Pyspark and Kafka"
authors = ["mDran <dran2112@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyspark = "3.3.0"
findspark = "2.0.1"
numpy = "^1.25.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pylint = "^2.17.4"
black = "^23.3.0"
kafka-python = "2.0.2"
jupyterlab = "^4.0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
